Fix for potential null fields in mixed modality datasets (#2214)
The current logic in `JsonlCollectionIterator` either expects all the fields to be present and non-None in the json line or be under the `contents` key (for the latter missing fields are replaced with ""). After this change all the fields present in the json line can have null values. This is required for M-BEIR and other future mixed modality dataset integrations.